Major refactoring
- create typedefs for different uses of linked_list and hash_table
- move function definitions so that purpose of files is more distinct
- use size_t more consistently for values that should not be negative

+++ entry.c
@@ -38,7 +38,7 @@ ioopm_entry_t *ioopm_find_previous_entry_for_key(ioopm_entry_t *entry,
   return entry;
 }
 
-int ioopm_entry_size(ioopm_entry_t *entry) {
+size_t ioopm_entry_size(ioopm_entry_t *entry) {
   if (entry == NULL) {
     return 0;
   } else {
